# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  device.CreateDeviceRequest:
    properties:
      is_location_public:
        example: false
        type: boolean
      is_outdoor:
        example: false
        type: boolean
      is_public:
        example: false
        type: boolean
      name:
        example: Living Room Sensor
        type: string
    required:
    - name
    type: object
  device.CreateDeviceResponse:
    properties:
      data:
        $ref: '#/definitions/device.CreatedDevice'
    type: object
  device.CreatedDevice:
    properties:
      created_at:
        type: string
      device_id:
        example: dev_9f1c8a3b4d2e4f0a5b6c7d8e9f001122
        type: string
      device_key:
        example: sk_1a2b3c4d5e6f...store this now, it won't be shown again
        type: string
      id:
        type: string
      is_location_public:
        example: false
        type: boolean
      is_outdoor:
        example: false
        type: boolean
      is_public:
        example: false
        type: boolean
      name:
        example: Living Room Sensor
        type: string
      updated_at:
        type: string
    type: object
  device.Device:
    properties:
      created_at:
        type: string
      device_id:
        example: dev_9f1c8a3b4d2e4f0a5b6c7d8e9f001122
        type: string
      id:
        type: string
      is_location_public:
        example: false
        type: boolean
      is_outdoor:
        example: false
        type: boolean
      is_public:
        example: false
        type: boolean
      name:
        example: Living Room Sensor
        type: string
      updated_at:
        type: string
    type: object
  device.ErrorResponse:
    properties:
      details:
        example: name is required
        type: string
      error:
        example: Invalid request body
        type: string
    type: object
  device.ListDevicesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/device.Device'
        type: array
    type: object
...
  device.PublicDevice:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_outdoor:
        type: boolean
      is_public:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  device.PublicDeviceResponse:
    properties:
      data:
        $ref: '#/definitions/device.PublicDevice'
    type: object
  device.RotateKeyResponse:
    properties:
      data:
        $ref: '#/definitions/device.CreatedDevice'
    type: object
  device.UpdateDeviceRequest:
    properties:
      is_location_public:
        type: boolean
      is_outdoor:
        type: boolean
      is_public:
        type: boolean
      name:
        example: Balcony Sensor
        minLength: 1
        type: string
    type: object
  device.UpdateDeviceResponse:
    properties:
      data:
        $ref: '#/definitions/device.Device'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Air quality monitoring API
  title: Air Quality API
  version: "1.0"
paths:
...
  /devices:
    get:
      description: Returns all registered devices. Secret keys are never included.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.ListDevicesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/device.ErrorResponse'
      summary: List all devices
      tags:
      - Devices
    post:
      consumes:
      - application/json
      description: Registers a device and returns its credentials. The device_key
        is a secret shown only once here; store it securely.
      parameters:
      - description: Device to register
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/device.CreateDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/device.CreateDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/device.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/device.ErrorResponse'
      summary: Register a new device
      tags:
      - Devices
  /devices/{id}:
    delete:
      description: Permanently deletes a device by its id.
      parameters:
      - description: Device id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/device.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/device.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/device.ErrorResponse'
      summary: Delete a device
      tags:
      - Devices
    patch:
      consumes:
      - application/json
      description: Partially updates a device. Exact location can only be shared by
        a public device, and making a device private revokes location sharing.
      parameters:
      - description: Device id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/device.UpdateDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.UpdateDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/device.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/device.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/device.ErrorResponse'
      summary: Update a device
      tags:
      - Devices
...
  /devices/{id}/rotate-key:
    post:
      description: Generates a new device_key for the device, invalidating the old
        one immediately. The new key is shown only in this response; update the physical
        device with it right away.
      parameters:
      - description: Device id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.RotateKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/device.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/device.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/device.ErrorResponse'
      summary: Rotate a device's secret key
      tags:
      - Devices
...
  /public/devices/{id}:
    get:
      parameters:
      - description: Device id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.PublicDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/device.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/device.ErrorResponse'
      summary: Get a public device
      tags:
      - Public
...
swagger: "2.0"